Update sbitx_gtk.c to restrict minimum screen height to 430 pixels
If the screen height is set to 425 or less pixels, then when the CW keyboard is opened, the sbitx application crashes. This occurs because the calculated screen height of the waterfall becomes a negative number. That height is used in a malloc() call near line 1523 in sbitx_gtk.c Obviously the malloc() call does not work well with a negative number, so it fails, and sbitx exits.
This change prevents the waterfall area from having a height less than 0, which in turn prevents the malloc() call from failing.