Update sbitx_gtk.c to improve handing of errors when attempting to write user settings
This change will cause the next attempt to write the user settings file when there is a file write error.

For example, if the user decides to write protect the user settings file to reduce wear on the sd card, the code will detect the error and return with the settings_updated flag cleared. However the last_save_at variable was not getting updated during the handling of this error, so the function could be called again immediately.